Migrate old store web app code to new version
