--- sample-temp/.editorconfig
@@ -1,9 +1,0 @@
-﻿[*.cs]
-
-# Sample Web App Effects rules
-[**/*Effects.cs]
-dotnet_diagnostic.SA1313.severity = none           # Parameter names should begin with lower-case letter
-
-# Sample Web App Reducer rules
-[**/*Reducer.cs]
-dotnet_diagnostic.SA1313.severity = none           # Parameter names should begin with lower-case letter
