--- sample-temp/Pizzeria.Store.Web/GlobalSuppressions.cs
@@ -1,6 +1,0 @@
-using System.Diagnostics.CodeAnalysis;
-
-[assembly: SuppressMessage(
-    "Performance",
-    "CA1515:Consider making public types internal",
-    Justification = "Public types are needed for bUnit testing")]
