--- sample-temp/Pizzeria.Store.Web/Layout/MainLayout.razor
@@ -1,24 +1,0 @@
-﻿@inherits LayoutComponentBase
-
-<MudThemeProvider />
-<MudPopoverProvider />
-<MudDialogProvider />
-<MudSnackbarProvider />
-
-<MudLayout>
-    <MudAppBar Elevation="1">
-        <MudIconButton Icon="Icons.Material.Filled.Menu" Color="Color.Inherit" Edge="Edge.Start" />
-        <MudSpacer />
-        <MudText Typo="Typo.h6">Lewee Pizzeria</MudText>
-        <MudSpacer />
-        <form method="post" action="@Routes.SignOut">
-            <AntiforgeryToken />
-            <MudIconButton Icon="@Icons.Material.Filled.Logout" aria-label="sign-out" Color="Color.Inherit" ButtonType="ButtonType.Submit" />
-        </form>
-    </MudAppBar>
-    <MudMainContent>
-        <MudContainer MaxWidth="MaxWidth.Large" Class="my-16 pt-16">
-            @Body
-        </MudContainer>
-    </MudMainContent>
-</MudLayout>
