--- sample-temp/Pizzeria.Store.Web/Layout/MainLayout.razor.cs
@@ -1,8 +1,0 @@
-using Microsoft.AspNetCore.Components;
-
-namespace Pizzeria.Store.Web.Layout;
-
-public partial class MainLayout : LayoutComponentBase
-{
-    public const string SignOutButtonSelector = "button[aria-label='sign-out']";
-}
