--- sample-temp/Pizzeria.Store.Web/Pages/Home.razor
@@ -1,72 +1,0 @@
-﻿@page "/"
-
-@using Pizzeria.Store.Web.States.Orders
-
-@attribute [Microsoft.AspNetCore.Authorization.Authorize]
-@inherits FluxorComponent
-
-@inject IState<OrdersState> State
-@inject IDispatcher Dispatcher
-@inject NavigationManager NavigationManager
-
-<PageTitle>Pizzeria Store</PageTitle>
-
-<MudText Typo="Typo.h3" Align="Align.Center" GutterBottom="true">
-    Welcome to Lewee Pizzeria
-</MudText>
-
-@if (this.State.Value.CurrentOrder == null)
-{
-    <MudGrid Justify="Justify.Center">
-        <MudItem xs="12" sm="6" md="4">
-            <MudCard>
-                <MudCardContent>
-                    <MudText Typo="Typo.h6" GutterBottom="true">
-                        Ready to Order?
-                    </MudText>
-                    <MudText Typo="Typo.body2">
-                        Start a new order to browse our delicious pizza menu and add your favorites to your cart.
-                    </MudText>
-                </MudCardContent>
-                <MudCardActions>
-                    <MudButton Variant="Variant.Filled" 
-                               Color="Color.Primary" 
-                               Size="Size.Large"
-                               FullWidth="true"
-                               OnClick="StartNewOrder"
-                               Disabled="this.State.Value.IsStartingOrder">
-                        @if (this.State.Value.IsStartingOrder)
-                        {
-                            <MudProgressCircular Size="Size.Small" Indeterminate="true" />
-                            <MudText Class="ml-2">Starting Order...</MudText>
-                        }
-                        else
-                        {
-                            <MudIcon Icon="Icons.Material.Filled.ShoppingCart" />
-                            <MudText Class="ml-2">Start New Order</MudText>
-                        }
-                    </MudButton>
-                </MudCardActions>
-            </MudCard>
-        </MudItem>
-    </MudGrid>
-}
-else
-{
-    <MudAlert Severity="Severity.Success" Class="mb-4">
-        <MudText Typo="Typo.body1">
-            Order started successfully! 
-            <MudLink Href="@Routes.Order" Color="Color.Primary">Continue to menu</MudLink>
-        </MudText>
-    </MudAlert>
-}
-
-@if (!string.IsNullOrEmpty(State.Value.ErrorMessage))
-{
-    <MudAlert Severity="Severity.Error" 
-              Class="mt-4" 
-              ShowCloseIcon="true" 
-              CloseIconClicked="ClearError">
-        @this.State.Value.ErrorMessage
-    </MudAlert>
-}
