--- sample-temp/Pizzeria.Store.Web/Pages/Home.razor.cs
@@ -1,16 +1,0 @@
-using Pizzeria.Store.Web.States.Orders.Actions;
-
-namespace Pizzeria.Store.Web.Pages;
-
-public partial class Home
-{
-    private void StartNewOrder()
-    {
-        this.Dispatcher.Dispatch(new StartOrderAction());
-    }
-
-    private void ClearError()
-    {
-        this.Dispatcher.Dispatch(new ClearOrderErrorAction());
-    }
-}
