--- sample-temp/Pizzeria.Store.Web/Pizzeria.Store.Web.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <ItemGroup>
-    <PackageReference Include="Aspire.Keycloak.Authentication" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.OpenIdConnect" />
-    <PackageReference Include="MudBlazor" />
-    <PackageReference Include="Fluxor.Blazor.Web" />
-    <PackageReference Include="Refit.HttpClientFactory" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\sample\Pizzeria.Common\Pizzeria.Common.csproj" />
-    <ProjectReference Include="..\..\sample\Pizzeria.ServiceDefaults\Pizzeria.ServiceDefaults.csproj" />
-    <ProjectReference Include="..\..\sample\Pizzeria.Store.Contracts\Pizzeria.Store.Contracts.csproj" />
-    <ProjectReference Include="..\..\src\Lewee.Blazor\Lewee.Blazor.csproj" />
-  </ItemGroup>
-
-</Project>
