--- sample-temp/Pizzeria.Store.Web/Routes.cs
@@ -1,8 +1,0 @@
-namespace Pizzeria.Store.Web;
-
-public static class Routes
-{
-    public const string Home = "/";
-    public const string SignOut = "/signout";
-    public const string Order = "/order";
-}
