--- sample-temp/Pizzeria.Store.Web/States/Orders/Actions/AddPizzaToOrderAction.cs
@@ -1,3 +1,0 @@
-namespace Pizzeria.Store.Web.States.Orders.Actions;
-
-public record AddPizzaToOrderAction(Guid OrderId, Guid PizzaId);
