--- sample-temp/Pizzeria.Store.Web/States/Orders/Actions/AddPizzaToOrderFailureAction.cs
@@ -1,3 +1,0 @@
-namespace Pizzeria.Store.Web.States.Orders.Actions;
-
-public record AddPizzaToOrderFailureAction(Guid PizzaId, string ErrorMessage);
