--- sample-temp/Pizzeria.Store.Web/States/Orders/Actions/StartOrderAction.cs
@@ -1,8 +1,0 @@
-using Lewee.Blazor.Fluxor.Actions;
-
-namespace Pizzeria.Store.Web.States.Orders.Actions;
-
-public record StartOrderAction : IRequestAction
-{
-    public Guid CorrelationId { get; init; } = Guid.NewGuid();
-}
