--- sample-temp/Pizzeria.Store.Web/States/Orders/Actions/StartOrderCompletedAction.cs
@@ -1,16 +1,0 @@
-using Lewee.Blazor.Fluxor.Actions;
-using Pizzeria.Store.Contracts;
-
-namespace Pizzeria.Store.Web.States.Orders.Actions;
-
-public record StartOrderCompletedAction : IMessageReceivedAction
-{
-    public StartOrderCompletedAction(OrderDto order, Guid correlationId)
-    {
-        this.Order = order;
-        this.CorrelationId = correlationId;
-    }
-
-    public OrderDto Order { get; init; }
-    public Guid CorrelationId { get; init; }
-}
