--- sample-temp/Pizzeria.Store.Web/States/Orders/Actions/StartOrderSuccessAction.cs
@@ -1,8 +1,0 @@
-using Lewee.Blazor.Fluxor.Actions;
-
-namespace Pizzeria.Store.Web.States.Orders.Actions;
-
-public record StartOrderSuccessAction : IRequestSuccessAction
-{
-    public Guid CorrelationId { get; init; }
-}
