--- sample-temp/Pizzeria.Store.Web/_Imports.razor
@@ -1,17 +1,0 @@
-﻿@using System.Net.Http
-@using System.Net.Http.Json
-@using Fluxor
-@using Fluxor.Blazor.Web.Components
-@using Microsoft.AspNetCore.Components.Forms
-@using Microsoft.AspNetCore.Components.Routing
-@using Microsoft.AspNetCore.Components.Web
-@using Microsoft.AspNetCore.Components.Web.HtmlRendering
-@using Microsoft.AspNetCore.Components.Web.Virtualization
-@using Microsoft.JSInterop
-@using MudBlazor
-@using Pizzeria.Common
-@using Pizzeria.Store.Contracts
-@using Pizzeria.Store.Web
-@using Pizzeria.Store.Web.Layout
-@using Pizzeria.Store.Web.Services
-@using static Microsoft.AspNetCore.Components.Web.RenderMode
