+++ sample/Pizzeria.Store.Web/Orders/Order.razor
@@ -1,5 +1,6 @@
-@page "/order"
+@attribute [Route(PageRoutes.Order)]
 @attribute [Microsoft.AspNetCore.Authorization.Authorize]
+
 @inherits FluxorComponent
 
 <PageTitle>Order - Pizzeria Store</PageTitle>
