+++ sample/Pizzeria.Store.Web/Pizzas/Actions/LoadPizzasAction.cs
@@ -1,9 +1,9 @@
 using System.Diagnostics.CodeAnalysis;
 
-namespace Pizzeria.Store.Web.States.Orders.Actions;
+namespace Pizzeria.Store.Web.Pizzas.Actions;
 
 [SuppressMessage(
     "Minor Code Smell",
     "S2094:Classes should not be empty",
     Justification = "False positive")]
-public record ClearOrderErrorAction;
+public record LoadPizzasAction;
