+++ sample/Pizzeria.Store.Web/Pizzas/Actions/LoadPizzasFailureAction.cs
@@ -1,3 +1,3 @@
-namespace Pizzeria.Store.Web.States.Pizzas.Actions;
+namespace Pizzeria.Store.Web.Pizzas.Actions;
 
 public record LoadPizzasFailureAction(string ErrorMessage);
