+++ sample/Pizzeria.Store.Web/Pizzas/PizzasEffects.cs
@@ -1,9 +1,9 @@
 using System.Diagnostics.CodeAnalysis;
 using Fluxor;
-using Pizzeria.Store.Web.Services;
-using Pizzeria.Store.Web.States.Pizzas.Actions;
+using Pizzeria.Store.Web.Infrastructure;
+using Pizzeria.Store.Web.Pizzas.Actions;
 
-namespace Pizzeria.Store.Web.States.Pizzas;
+namespace Pizzeria.Store.Web.Pizzas;
 
 public class PizzasEffects
 {
